driver: apdu: rename QRTR QMI helpers to QMI helpers
Almost all of the QRTR QMI helpers will be reused for direct QMI so,
lets rename the sources to indicate that.
