+++ driver/apdu/qmi_helpers.h
